12/7 setup event based public scripts
