+++ public/index.html
@@ -45,8 +45,8 @@
           <button id="copyRoomIdElement" class="btn btn-outline-primary  btn-sm">Copy Room ID</button>
         </div>
         <div id="playersAndVotes">
-          <ul id="socketsInTheRoom">Players:</ul>
-          <ul id="socketsInTheRoomVoteNumbers">Votes:</ul>
+          <ul id="socketsInTheRoomElement">Players:</ul>
+          <ul id="socketsInTheRoomVoteNumbersElement">Votes:</ul>
         </div>
     
       <div id="vote-area">
@@ -61,15 +61,12 @@
       </form>
       </div>
 
-      <!-- <button id="testButton" value="testButton" class="btn btn-secondary">Test</button> -->
-      <!-- if i delete testButton it breaks the app wtf -->
-      <!-- <button id="nextRoundButton"  class="btn btn-dark">Next Round</button> -->
       <button id="showVotesButton" value="showVotesButtonValue" class="btn btn-dark">Show Votes</button>
 
       <a href="" onclick="changePage('page1')" class="btn btn-light" id="leaveRoomButton">Leave room</a>
       <!-- TODO: switch onclick to addEventListener -->
       <div>
-        <p id="average"></p>
+        <p id="averageElement"></p>
       </div>
